requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
albumentations>=1.3.0
scikit-learn>=1.2.0
scikit-image>=0.21.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
facenet-pytorch>=2.5.0
opencv-python>=4.8.0
